Refactor Redis server implementation
- Updated .gitignore to exclude additional IDE and build files.
- Enhanced VSCode settings for better C++ support with file associations and clangd configuration.
- Modified CMakeLists.txt to export compile commands and reorganize source files for better structure.
- Removed obsolete Handler class and its implementation, replacing it with a more modular request handling system.
- Introduced RequestExecutor to manage client requests concurrently.
- Created ServerApp to encapsulate server logic, including connection handling and request processing.
- Implemented various request handlers (Get, Set, Del, Unknown) to process specific commands.
- Added concurrent queue for managing client file descriptors.
- Updated RespValue and RespValueQueue classes for improved parsing and handling of RESP protocol.
- Cleaned up and organized header files for better readability and maintainability.
- Added utility functions for sending responses and handling errors.

+++ src/respvalue/RespValue.cpp
@@ -1,7 +1,6 @@
 #include "RespValue.h"
 
 #include <cassert>
-#include <sstream>
 #include <stdexcept>
 
 RespValue::RespValue(RespVariant variant) : value_(std::move(variant)) {}
@@ -83,8 +82,7 @@ RespValue::RespBulkString RespValue::parseBulkString(const std::string& str,
   }
   // Ensure there's enough data for the bulk string content plus trailing CRLF.
   if (bulkStringLength < 0 ||
-      (static_cast<size_t>(pos) + static_cast<size_t>(bulkStringLength) + 2) >
-          str.size()) {
+      (pos + static_cast<size_t>(bulkStringLength) + 2) > str.size()) {
     throw std::out_of_range("Bulk string payload truncated or missing CRLF");
   }
   // Verify terminating CRLF after payload.
@@ -92,7 +90,8 @@ RespValue::RespBulkString RespValue::parseBulkString(const std::string& str,
       str[pos + bulkStringLength + 1] != '\n') {
     throw std::out_of_range("Bulk string missing terminating CRLF");
   }
-  std::string bulkString = str.substr(pos, static_cast<size_t>(bulkStringLength));
+  std::string bulkString =
+      str.substr(pos, static_cast<size_t>(bulkStringLength));
   pos += static_cast<size_t>(bulkStringLength) + 2;
   return bulkString;
 }
